+++ src/js/api/posts/postByID.js
@@ -1,14 +1,16 @@
 // import { dateSetup } from "../../src/js/modules/source.mjs";
-import { API_SOSIAL_URL } from "../../src/js/api/constant-api.mjs";
-import { postContentID } from "../../src/js/modules/source.mjs";
-import { postData } from "../../src/js/api/posts/create.mjs";
+import { API_SOSIAL_URL } from "../constant-api.mjs";
+import { postContentID } from "../../modules/source.mjs";
+import { editPostContentForm } from "../../modules/source.mjs";
+import { loadComments } from "./comments/loadComments.mjs";
+import { postData } from "./apiCall.mjs";
 
 const queryString = document.location.search;
 const params = new URLSearchParams(queryString);
 const productId = params.get("id");
 
-const action = "/posts";
 const method = "GET";
+const action = "/posts";
 const postAuthor = "?_author=true";
 const postIdUrl = `${API_SOSIAL_URL}${action}/${productId}${postAuthor}`;
 
@@ -21,7 +23,15 @@ const postIdUrl = `${API_SOSIAL_URL}${action}/${productId}${postAuthor}`;
 async function loadPosts(postIdUrl, method, data) {
   try {
     const postInfo = await postData(postIdUrl, method, data);
-    postContentID(postInfo);
+    // postContentID(postInfo);
+
+    const path = location.pathname;
+    if (path.includes("/profile/editPosts.html")) {
+      editPostContentForm(postInfo);
+      loadComments;
+    } else if (path.includes("/feed/specific-post.html")) {
+      postContentID(postInfo);
+    }
   } catch (error) {
     console.log(error);
   }
