+++ src/js/api/posts/userPosts.js
@@ -1,13 +1,13 @@
 // import { dateSetup } from "../../src/js/modules/source.mjs";
-import { API_SOSIAL_URL } from "../../src/js/api/constant-api.mjs";
-import { postData } from "../../src/js/api/posts/create.mjs";
-import { getLocalStorage } from "../../src/js/storage/index.mjs";
-import { userPostsContent } from "../../src/js/modules/source.mjs";
+import { API_SOSIAL_URL } from "../constant-api.mjs";
+import { postData } from "./apiCall.mjs";
+import { getLocalStorage } from "../../localStorage/index.mjs";
+import { userPostsContent } from "../../modules/source.mjs";
 
-const action = "/profiles";
 const user = getLocalStorage("profile").name;
-const postTrue = "?_posts=true";
 const method = "GET";
+const action = "/profiles";
+const postTrue = "?_posts=true";
 const profileUrl = `${API_SOSIAL_URL}${action}/${user}${postTrue}&_count=true`;
 
 /**
@@ -20,7 +20,10 @@ const profileUrl = `${API_SOSIAL_URL}${action}/${user}${postTrue}&_count=true`;
 async function loadPosts(profileUrl, method, data) {
   try {
     const postInfo = await postData(profileUrl, method, data);
-    Object.values(postInfo.posts).forEach(function (post) {
+    const userPosts = postInfo.posts;
+
+    const newPostInfo = userPosts.sort((a, b) => new Date(b.updated) - new Date(a.updated));
+    Object.values(newPostInfo).forEach(function (post) {
       userPostsContent(post);
     });
   } catch (error) {
