+++ src/js/filterSearch/filter.js
@@ -1,6 +1,6 @@
-import { API_SOSIAL_URL } from "../../src/js/api/constant-api.mjs";
-import { postData } from "../../src/js/api/posts/create.mjs";
-import { postContent } from "../../src/js/modules/source.mjs";
+import { API_SOSIAL_URL } from "../api/constant-api.mjs";
+import { postData } from "../api/posts/apiCall.mjs";
+import { resultFeedUpdate } from "../modules/source.mjs";
 
 const popularPosts = document.querySelector("#popular-posts");
 const newPosts = document.querySelector("#new-posts");
@@ -20,8 +20,8 @@ const postUrl = `${API_SOSIAL_URL}${action}${postAuthor}`;
 async function popularFilter() {
   try {
     const postInfo = await postData(postUrl, method);
-    const filterPosts = postInfo.sort((a, b) => b._count.reactions - a._count.reactions);
-    resultFeedUpdate(filterPosts);
+    const result = postInfo.sort((a, b) => b._count.reactions - a._count.reactions);
+    resultFeedUpdate(result);
   } catch (error) {
     console.log(error);
   }
@@ -48,21 +48,6 @@ async function oldPostFilter() {
     const postInfo = await postData(postUrl);
     const oldPostDate = postInfo.sort((a, b) => new Date(a.updated) - new Date(b.updated));
     resultFeedUpdate(oldPostDate);
-  } catch (error) {
-    console.log(error);
-  }
-}
-
-/**
- * This function shows the result in the feed page
- * @param {string} filterPosts The result of the sort (popular/new/old) from one of the other functions
- */
-function resultFeedUpdate(filterPosts) {
-  try {
-    userPost.textContent = "";
-    Object.values(filterPosts).forEach(function (post) {
-      postContent(post);
-    });
   } catch (error) {
     console.log(error);
   }
